AGATHA-722: Increase likelihood of reading CC values phy driver.
Previously we had an issue where we missed new values on the CC
pins during USB power negotiation. We tried to fix this by calling
the negotiation function from the bifrost_rpmsg driver. However
this caused an issue where the negotiation failed, and the fix
for that issue in turn caused a kernel panic during resume after
suspend.

This new code tries to solve all of those problems by making
the phy driver wait for a while for new CC pin values. There's
still a chance that we won't read these values if it takes too long.

Signed-off-by: Anton Persson <anton.persson@teledyneflir.com>